packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minions-prompts"
version = "0.1.0"
description = "Version-controlled prompt engineering system built on the Minions SDK"
readme = "README.md"
license = { text = "AGPL-3.0" }
requires-python = ">=3.11"
keywords = ["ai", "prompts", "minions", "versioning", "llm"]
authors = [
    { name = "Mehdi Nabhani", email = "mehdi@the-mehdi.com" }
]
dependencies = [
    "minions-sdk>=0.2.0",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/mxn2020/minions-prompts"
Repository = "https://github.com/mxn2020/minions-prompts"

[tool.hatch.build.targets.wheel]
packages = ["minions_prompts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
